Support force_frac parameter in format_percent
This adds a force_frac parameter in format_percent method which allows the user to override the fractional precision of the outputed number.  This is error prone and costly to do by directly overriding the format string.  To do so assumes the shape of the default format string for the locale and current character encoding. This may be an incorrect assumption if there is a space between the percent sign and the rest of the number.

This also adds tests and documentation for the new parameter.